Avoid circular import differently
